+++ Packages/com.texelsaur.video/Runtime/Scripts/Component/Audio/AudioChannel.cs
@@ -15,6 +15,8 @@ namespace Texel
         FOUR,
         FIVE,
         SIX,
+        SEVEN,
+        EIGHT,
     }
 
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]